Move IRQ related structure and functions into kvm/irq.go
